+++ data/eventsData.js
@@ -1,9 +1,14 @@
 export const eventsArray = [
-    {
-        _id: 0,
-        date: new Date('2023-01-23').toDateString(),
-        imgs: ['C:/Users/akino/coding_projects/21ctl/events/assets/img/events-1', 'C:/Users/akino/coding_projects/21ctl/events/assets/img/events-2'],
-        title: 'Digital Experts Academy Launch',
-        body: 'On Monday January 23, 2023, 21st Century Technologies, led by the visionairy chairman Mr. Wale Ajisebutu, unveiled the Digital Experts Academy, which was declared open by Mr. Will Stevens, the United States Consul General, in the presence of Mr. Wole Adeniyi, the Chief Executive Officer (CEO) of Stanbic IBTC Bank, Dr. Dare Obadare, the MD/Chairman of Digital Encode and notable top executives from Oracle, Microsoft, VMWare.'
-    }
-]
+  {
+    _id: "1",
+    date: new Date('2023-01-23').toDateString(),
+    imgs: [
+      'https://res.cloudinary.com/dqsggbqmf/image/upload/v1675250342/21ctl/1674495588828_bpf9lm.jpg',
+      'https://res.cloudinary.com/dqsggbqmf/image/upload/v1675250386/21ctl/1674495589326_yyqqyd.jpg',
+      'https://res.cloudinary.com/dqsggbqmf/image/upload/v1675250411/21ctl/1674495589157_or9oaf.jpg',
+      'https://res.cloudinary.com/dqsggbqmf/image/upload/v1675250442/21ctl/1674495588974_jb8krp.jpg',
+    ],
+    title: ' Unveiling the Digital Experts Academy (DEA)',
+    body: 'On Monday January 23, 2023, 21st Century Technologies, led by the visionairy chairman Mr. Wale Ajisebutu, unveiled the Digital Experts Academy, which was declared open by Mr. Will Stevens, the United States Consul General, in the presence of Mr. Wole Adeniyi, the Chief Executive Officer (CEO) of Stanbic IBTC Bank, Dr. Dare Obadare, the MD/Chairman of Digital Encode and notable top executives from Oracle, Microsoft, VMWare.',
+  },
+];
