+++ src/components/layout/index.js
@@ -2,4 +2,5 @@ export * from './LightNavbar';
 export * from './Footer';
 export * from './SocialIcons';
 export * from './DigitalNavbar'
+export * from './EventLayout'
